requirements.txt
torch
transformers
langchain
sentence-transformers
faiss-cpu         # or faiss-gpu if using a GPU
datasets
